Adiciona cadastro de livros

+++ src/main/java/com/zup/cqc/repository/CategoryRepository.java
@@ -1,7 +1,7 @@
 package com.zup.cqc.repository;
 
-import com.zup.cqc.category.Category;
+import com.zup.cqc.entity.category.Category;
 import org.springframework.data.repository.CrudRepository;
 
-public interface CategoryRepository extends CrudRepository<Category, Integer> {
+public interface CategoryRepository extends CrudRepository<Category, Long> {
 }
